requirements.txt
pytubefix
google-generativeai
playwright
shutil
smtplib
email
requests
